Refactor bindings
