--- extensions/bindings/cs/net.csproj
@@ -1,61 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Project DefaultTargets="Build" ToolsVersion="3.5" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProductVersion>9.0.21022</ProductVersion>
-    <SchemaVersion>2.0</SchemaVersion>
-    <ProjectGuid>{2F7BA85D-3805-4FF8-B666-9D1235BBA14B}</ProjectGuid>
-    <OutputType>Library</OutputType>
-    <RootNamespace>net</RootNamespace>
-    <AssemblyName>net</AssemblyName>
-    <TargetFrameworkVersion>v3.5</TargetFrameworkVersion>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-    <DebugSymbols>true</DebugSymbols>
-    <DebugType>full</DebugType>
-    <Optimize>false</Optimize>
-    <OutputPath>bin\Debug</OutputPath>
-    <DefineConstants>DEBUG</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-    <ConsolePause>false</ConsolePause>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <DebugType>none</DebugType>
-    <Optimize>false</Optimize>
-    <OutputPath>bin\Release</OutputPath>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-    <ConsolePause>false</ConsolePause>
-  </PropertyGroup>
-  <ItemGroup>
-    <Reference Include="System" />
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Include="AssemblyInfo.cs" />
-    <Compile Include="Network.cs" />
-    <Compile Include="net_internal.cs" />
-    <Compile Include="net_internalPINVOKE.cs" />
-    <Compile Include="Sugar.cs" />
-    <Compile Include="LogLevel.cs" />
-    <Compile Include="Node.cs" />
-    <Compile Include="NTA_BasicType.cs" />
-    <Compile Include="NTA_LogLevel.cs" />
-    <Compile Include="NTA_Node.cs" />
-    <Compile Include="Region.cs" />
-    <Compile Include="SWIGTYPE_p_Node.cs" />
-    <Compile Include="SWIGTYPE_p_std__vectorTint_t.cs" />
-    <Compile Include="SWIGTYPE_p_std__vectorTstd__string_t.cs" />
-    <Compile Include="LoggingException.cs" />
-    <Compile Include="LogItem.cs" />
-    <Compile Include="SWIGTYPE_p_Byte.cs" />
-    <Compile Include="SWIGTYPE_p_std__ostringstream.cs" />
-    <Compile Include="SWIGTYPE_p_UInt32.cs" />
-    <Compile Include="BaseException.cs" />
-  </ItemGroup>
-  <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
-  <ItemGroup>
-    <None Include="net_internal.i" />
-  </ItemGroup>
-</Project>
