--- extensions/bindings/cs/net.dll.config
@@ -1,3 +1,0 @@
- <configuration>
-   <dllmap dll="net_internal" target="net_internal.so" />
- </configuration>
