--- extensions/bindings/cs/net.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 10.00
-# Visual Studio 2008
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "net", "net.csproj", "{2F7BA85D-3805-4FF8-B666-9D1235BBA14B}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "net_test", "net_test\net_test.csproj", "{AA6C6216-F9B6-49B2-99C8-9F4E56F049A3}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-		Debug|x86 = Debug|x86
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{2F7BA85D-3805-4FF8-B666-9D1235BBA14B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2F7BA85D-3805-4FF8-B666-9D1235BBA14B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2F7BA85D-3805-4FF8-B666-9D1235BBA14B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2F7BA85D-3805-4FF8-B666-9D1235BBA14B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{AA6C6216-F9B6-49B2-99C8-9F4E56F049A3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AA6C6216-F9B6-49B2-99C8-9F4E56F049A3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AA6C6216-F9B6-49B2-99C8-9F4E56F049A3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AA6C6216-F9B6-49B2-99C8-9F4E56F049A3}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(MonoDevelopProperties) = preSolution
-		StartupItem = net_test\net_test.csproj
-	EndGlobalSection
-EndGlobal
