+++ nupic/bindings/sparse_matrix.i
